backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
snowflake-connector-python
python-dotenv
sse-starlette
pydantic
